chore(test/lint-stage): add yarn test to pre-commmit
I was running into an issue where lint-staged would try and run tests, but if there are no tests
file present, it would fail. Now, the testing is done via husky in the pre-commit hook. As a result,
if there are no test files that are staged, lint-stage will not fail, and we con continue the
commit.